Fixed bug in ShortGenerator.  It now returns a masked short which much be <= 2^15.

+++ headers/Socket.h
@@ -107,7 +107,7 @@ private:
     int domain_;
     int type_;
     int protocol_;
-    int16_t socket_;
+    int socket_;
 };
 
 #endif	/* _SOCKET_H */
